+++ src/main/java/module-info.java
@@ -20,10 +20,10 @@
  */
 module io.github.astrapisixtynine.silly.collection
 {
+	requires io.github.astrapisixtynine.comparator.extensions;
 	requires io.github.astrapisixtynine.jobj.core;
+	requires io.github.astrapisixtynine.silly.bean;
 	requires org.apache.commons.collections4;
-	requires comparator.extensions.main;
-	requires silly.bean.main;
 
 	exports io.github.astrapi69.collection;
 	exports io.github.astrapi69.collection.array;