+++ src/test/java/io/github/astrapi69/collection/migrate/MigrateToTomlVersions.java
@@ -26,12 +26,9 @@ package io.github.astrapi69.collection.migrate;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.List;
-import java.util.Map;
 
 import org.junit.jupiter.api.Test;
 
-import io.github.astrapi69.file.read.ReadFileExtensions;
 import io.github.astrapi69.file.search.PathFinder;
 
 /**